Renamed email_prefixer to neeto-email-prefixer
