+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in email_prefixer.gemspec
+# Specify your gem's dependencies in neeto-email-prefixer.gemspec
 gemspec
