--- lib/email_prefixer/version.rb
@@ -1,3 +1,0 @@
-module EmailPrefixer
-  VERSION = '1.2.0'
-end
