+++ lib/neeto-email-prefixer.rb
@@ -1,0 +1,17 @@
+require 'neeto_email_prefixer/version'
+require 'neeto_email_prefixer/configuration'
+require 'neeto_email_prefixer/interceptor'
+require 'neeto_email_prefixer/railtie' if defined?(Rails)
+
+module NeetoEmailPrefixer
+  class << self
+    def configure
+      yield configuration
+      configuration
+    end
+
+    def configuration
+      @configuration ||= NeetoEmailPrefixer::Configuration.new
+    end
+  end
+end
