+++ lib/neeto_email_prefixer/railtie.rb
@@ -1,12 +1,12 @@
-module EmailPrefixer
+module NeetoEmailPrefixer
   class Railtie < ::Rails::Railtie
-    initializer 'email_prefixer.configure_defaults' do |app|
-      config = EmailPrefixer.configure do |config|
+    initializer 'neeto_email_prefixer.configure_defaults' do |app|
+      config = NeetoEmailPrefixer.configure do |config|
         config.application_name ||= app.class.module_parent_name
         config.stage_name ||= Rails.env
-        config.builder ||= EmailPrefixer::Configuration::DEFAULT_BUILDER
+        config.builder ||= NeetoEmailPrefixer::Configuration::DEFAULT_BUILDER
       end
-      interceptor = EmailPrefixer::Interceptor.new(config)
+      interceptor = NeetoEmailPrefixer::Interceptor.new(config)
       ActionMailer::Base.register_preview_interceptor(interceptor)
       ActionMailer::Base.register_interceptor(interceptor)
     end
