+++ lib/neeto_email_prefixer/version.rb
@@ -1,0 +1,3 @@
+module NeetoEmailPrefixer
+  VERSION = '1.2.0'
+end
